% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stacked_density.R
\name{plot_stacked_density}
\alias{plot_stacked_density}
\title{Create a stacked stacked density chart}
\usage{
plot_stacked_density(x, grp, xlim, xlab, ylab, y_labels = NULL)
}
\arguments{
\item{x}{values for the density plot.}

\item{grp}{variable containing the membership to groups.}

\item{xlab}{label of the x-axis.}

\item{ylab}{label of the y-axis.}

\item{y_labels}{labels at the ticks of the y-axis.}
}
\value{
A stacked density chart showing the density of \code{x} in the different groups contained in \code{grp}
}
\description{
Create a stacked stacked density chart
}
