% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_values.R
\name{recode_values}
\alias{recode_values}
\title{Recode the values in a vector}
\usage{
recode_values(x, from, to)
}
\arguments{
\item{x}{vector containing values}

\item{from}{vector containing the unique elements of x}

\item{to}{values in which the corresponging elements of from needs to be changed in}
}
\value{
Returns a vector where the values of x are changed from the values in the vector from to the values in the vector to
}
\description{
Recode the values in a vector
}
